+++ comet_pqc/measurements/cv_ramp.py
@@ -15,11 +15,11 @@ from ..benchmark import Benchmark
 
 from .matrix import MatrixMeasurement
 from .measurement import ComplianceError
-from .measurement import HVSourceMixin
-from .measurement import LCRMixin
-from .measurement import EnvironmentMixin
 from .measurement import format_estimate
 from .measurement import QUICK_RAMP_DELAY
+from .mixins import HVSourceMixin
+from .mixins import LCRMixin
+from .mixins import EnvironmentMixin
 
 __all__ = ["CVRampMeasurement"]
