+++ comet_pqc/measurements/iv_ramp.py
@@ -14,9 +14,9 @@ from ..formatter import PQCFormatter
 from ..estimate import Estimate
 from .matrix import MatrixMeasurement
 from .measurement import ComplianceError
-from .measurement import EnvironmentMixin
 from .measurement import format_estimate
 from .measurement import QUICK_RAMP_DELAY
+from .mixins import EnvironmentMixin
 
 __all__ = ["IVRampMeasurement"]
