+++ comet_pqc/panels/cv_ramp.py
@@ -5,9 +5,9 @@ from comet import ui
 
 from ..utils import format_metric
 from .matrix import MatrixPanel
-from .panel import HVSourceMixin
-from .panel import LCRMixin
-from .panel import EnvironmentMixin
+from .mixins import HVSourceMixin
+from .mixins import LCRMixin
+from .mixins import EnvironmentMixin
 
 __all__ = ["CVRampPanel"]
