+++ comet_pqc/panels/cv_ramp_alt.py
@@ -5,8 +5,8 @@ from comet import ui
 
 from ..utils import format_metric
 from .matrix import MatrixPanel
-from .panel import LCRMixin
-from .panel import EnvironmentMixin
+from .mixins import LCRMixin
+from .mixins import EnvironmentMixin
 
 __all__ = ["CVRampAltPanel"]
