+++ comet_pqc/panels/iv_ramp.py
@@ -6,8 +6,8 @@ from comet import ui
 
 from ..utils import format_metric
 from .matrix import MatrixPanel
-from .panel import HVSourceMixin
-from .panel import EnvironmentMixin
+from .mixins import HVSourceMixin
+from .mixins import EnvironmentMixin
 
 __all__ = ["IVRampPanel"]
