+++ comet_pqc/panels/iv_ramp_4_wire.py
@@ -3,8 +3,8 @@ from comet import ui
 
 from ..utils import format_metric
 from .matrix import MatrixPanel
-from .panel import VSourceMixin
-from .panel import EnvironmentMixin
+from .mixins import VSourceMixin
+from .mixins import EnvironmentMixin
 
 __all__ = ["IVRamp4WirePanel"]
