+++ comet_pqc/panels/iv_ramp_bias_elm.py
@@ -3,10 +3,10 @@ from comet import ui
 
 from ..utils import format_metric
 from .matrix import MatrixPanel
-from .panel import HVSourceMixin
-from .panel import VSourceMixin
-from .panel import ElectrometerMixin
-from .panel import EnvironmentMixin
+from .mixins import HVSourceMixin
+from .mixins import VSourceMixin
+from .mixins import ElectrometerMixin
+from .mixins import EnvironmentMixin
 
 __all__ = ["IVRampBiasElmPanel"]
