+++ comet_pqc/panels/iv_ramp_elm.py
@@ -5,9 +5,9 @@ from comet import ui
 
 from ..utils import format_metric
 from .matrix import MatrixPanel
-from .panel import HVSourceMixin
-from .panel import ElectrometerMixin
-from .panel import EnvironmentMixin
+from .mixins import HVSourceMixin
+from .mixins import ElectrometerMixin
+from .mixins import EnvironmentMixin
 
 __all__ = ["IVRampElmPanel"]
